backend/requirements/base.txt
# Master branches
https://github.com/django/django/archive/main.tar.gz
https://github.com/encode/django-rest-framework/archive/master.tar.gz

whitenoise==5.2.0

django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-phonenumber-field==5.1.0
djangorestframework-camel-case2==0.3.0

phonenumbers==8.12.21
